VU: Improve VU0/EE sync, Implement better M-Bit Handling, Fix VU program handing on VIF

+++ pcsx2/x86/microVU_Branch.inl
@@ -70,12 +70,36 @@ void mVUDTendProgram(mV, microFlagCycles* mFC, int isEbit) {
 		xMOVSS(ptr32[&mVU.regs().VI[REG_P].UL], xmmPQ);
 	}
 
-	// Save Flag Instances
-	xMOV(ptr32[&mVU.regs().VI[REG_STATUS_FLAG].UL],	getFlagReg(fStatus));
+	// Save MAC, Status and CLIP Flag Instances
+	xMOV(ptr32[&mVU.regs().VI[REG_STATUS_FLAG].UL], getFlagReg(fStatus));
 	mVUallocMFLAGa(mVU, gprT1, fMac);
 	mVUallocCFLAGa(mVU, gprT2, fClip);
-	xMOV(ptr32[&mVU.regs().VI[REG_MAC_FLAG].UL],	gprT1);
-	xMOV(ptr32[&mVU.regs().VI[REG_CLIP_FLAG].UL],	gprT2);
+	xMOV(ptr32[&mVU.regs().VI[REG_MAC_FLAG].UL], gprT1);
+	xMOV(ptr32[&mVU.regs().VI[REG_CLIP_FLAG].UL], gprT2);
+
+	if (!isEbit) { // Backup flag instances
+		xMOVAPS(xmmT1, ptr128[mVU.macFlag]);
+		xMOVAPS(ptr128[&mVU.regs().micro_macflags], xmmT1);
+		xMOVAPS(xmmT1, ptr128[mVU.clipFlag]);
+		xMOVAPS(ptr128[&mVU.regs().micro_clipflags], xmmT1);
+
+		xMOV(ptr32[&mVU.regs().micro_statusflags[0]], gprF0);
+		xMOV(ptr32[&mVU.regs().micro_statusflags[1]], gprF1);
+		xMOV(ptr32[&mVU.regs().micro_statusflags[2]], gprF2);
+		xMOV(ptr32[&mVU.regs().micro_statusflags[3]], gprF3);
+	} else { // Flush flag instances
+		xMOVDZX(xmmT1, ptr32[&mVU.regs().VI[REG_CLIP_FLAG].UL]);
+		xSHUF.PS(xmmT1, xmmT1, 0);
+		xMOVAPS(ptr128[&mVU.regs().micro_clipflags], xmmT1);
+
+		xMOVDZX(xmmT1, ptr32[&mVU.regs().VI[REG_MAC_FLAG].UL]);
+		xSHUF.PS(xmmT1, xmmT1, 0);
+		xMOVAPS(ptr128[&mVU.regs().micro_macflags], xmmT1);
+
+		xMOVDZX(xmmT1, ptr32[&mVU.regs().VI[REG_STATUS_FLAG].UL]);
+		xSHUF.PS(xmmT1, xmmT1, 0);
+		xMOVAPS(ptr128[&mVU.regs().micro_statusflags], xmmT1);
+	}
 
 	if (isEbit || isVU1) { // Clear 'is busy' Flags
 		if (!mVU.index || !THREAD_VU1) {
@@ -98,7 +122,12 @@ void mVUendProgram(mV, microFlagCycles* mFC, int isEbit) {
 	int fClip	= getLastFlagInst(mVUpBlock->pState, mFC->xClip,   2, isEbit);
 	int qInst	= 0;
 	int pInst	= 0;
-	mVU.regAlloc->flushAll();
+	microBlock stateBackup;
+	memcpy(&stateBackup, &mVUregs, sizeof(mVUregs)); //backup the state, it's about to get screwed with.
+	if(!isEbit)
+		mVU.regAlloc->TDwritebackAll(); //Writing back ok, invalidating early kills the rec, so don't do it :P
+	else
+		mVU.regAlloc->flushAll();
 
 	if (isEbit) {
 		memzero(mVUinfo);
@@ -124,19 +153,54 @@ void mVUendProgram(mV, microFlagCycles* mFC, int isEbit) {
 	}
 
 	// Save P/Q Regs
-	if (qInst) { xPSHUF.D(xmmPQ, xmmPQ, 0xe5); }
+	if (qInst) { xPSHUF.D(xmmPQ, xmmPQ, 0xe1); }
 	xMOVSS(ptr32[&mVU.regs().VI[REG_Q].UL], xmmPQ);
+	xPSHUF.D(xmmPQ, xmmPQ, 0xe1);
+	xMOVSS(ptr32[&mVU.regs().pending_q], xmmPQ);
+	xPSHUF.D(xmmPQ, xmmPQ, 0xe1);
+
 	if (isVU1) {
-		xPSHUF.D(xmmPQ, xmmPQ, pInst ? 3 : 2);
+		xPSHUF.D(xmmPQ, xmmPQ, pInst ? 0x1b : 0x1e);
 		xMOVSS(ptr32[&mVU.regs().VI[REG_P].UL], xmmPQ);
+		xPSHUF.D(xmmPQ, xmmPQ, pInst ? 0x1b : 0x4b);
+
+		xPSHUF.D(xmmPQ, xmmPQ, 0xe1);
+		xMOVSS(ptr32[&mVU.regs().pending_p], xmmPQ);
+		xPSHUF.D(xmmPQ, xmmPQ, 0x1b);
 	}
 
-	// Save Flag Instances
+	// Save MAC, Status and CLIP Flag Instances
 	xMOV(ptr32[&mVU.regs().VI[REG_STATUS_FLAG].UL],	getFlagReg(fStatus));
 	mVUallocMFLAGa(mVU, gprT1, fMac);
 	mVUallocCFLAGa(mVU, gprT2, fClip);
 	xMOV(ptr32[&mVU.regs().VI[REG_MAC_FLAG].UL],	gprT1);
 	xMOV(ptr32[&mVU.regs().VI[REG_CLIP_FLAG].UL],	gprT2);
+
+	if (!isEbit) { // Backup flag instances
+		xMOVAPS(xmmT1, ptr128[mVU.macFlag]);
+		xMOVAPS(ptr128[&mVU.regs().micro_macflags], xmmT1);
+		xMOVAPS(xmmT1, ptr128[mVU.clipFlag]);
+		xMOVAPS(ptr128[&mVU.regs().micro_clipflags], xmmT1);
+
+		xMOV(ptr32[&mVU.regs().micro_statusflags[0]], gprF0);
+		xMOV(ptr32[&mVU.regs().micro_statusflags[1]], gprF1);
+		xMOV(ptr32[&mVU.regs().micro_statusflags[2]], gprF2);
+		xMOV(ptr32[&mVU.regs().micro_statusflags[3]], gprF3);
+	}
+	else { // Flush flag instances
+		xMOVDZX(xmmT1, ptr32[&mVU.regs().VI[REG_CLIP_FLAG].UL]);
+		xSHUF.PS(xmmT1, xmmT1, 0);
+		xMOVAPS(ptr128[&mVU.regs().micro_clipflags], xmmT1);
+
+		xMOVDZX(xmmT1, ptr32[&mVU.regs().VI[REG_MAC_FLAG].UL]);
+		xSHUF.PS(xmmT1, xmmT1, 0);
+		xMOVAPS(ptr128[&mVU.regs().micro_macflags], xmmT1);
+
+		xMOVDZX(xmmT1, ptr32[&mVU.regs().VI[REG_STATUS_FLAG].UL]);
+		xSHUF.PS(xmmT1, xmmT1, 0);
+		xMOVAPS(ptr128[&mVU.regs().micro_statusflags], xmmT1);
+	}
+
 
 	if (isEbit || isVU1) { // Clear 'is busy' Flags
 		if (!mVU.index || !THREAD_VU1) {
@@ -149,6 +213,7 @@ void mVUendProgram(mV, microFlagCycles* mFC, int isEbit) {
 		xMOV(ptr32[&mVU.regs().VI[REG_TPC].UL], xPC);
 		xJMP(mVU.exitFunct);
 	}
+	memcpy(&mVUregs, &stateBackup, sizeof(mVUregs)); //Restore the state for the rest of the recompile
 }
 
 // Recompiles Code for Proper Flags and Q/P regs on Block Linkings
